hero section updated
